Remove `annoy` from pyright exclude
