+++ pyrightconfig.stricter.json
@@ -22,7 +22,6 @@
         "stdlib/xml/dom/pulldom.pyi",
         "stdlib/xml/sax",
         "stubs/aiofiles/aiofiles/threadpool/utils.pyi",
-        "stubs/annoy/annoy/annoylib.pyi",
         "stubs/aws-xray-sdk",
         "stubs/babel",
         "stubs/backports.ssl_match_hostname",
